Move AmqpConnectionSettings to its own file
